Move the C# body slicer out of ILInspector.Metadata
The metadata layer should expose PDB resources and nothing else, but
SourceLinkResolver.cs carried ~1,800 lines of C# text processing alongside
~420 lines of genuine SRM/PDB work. This moves the text half into a new
DotnetInspector.CSharpBodySlicer library and leaves the PDB half behind.

The two halves shared a file and nothing else, which is why the move is
mechanical:

  text half (lines 80-1906) -> SRM types      0 references
  SRM half -> text-half helpers               0 calls
  anything in Metadata -> ExtractMethodBody   none
  actual consumers                            Services and the CLI, both
                                              above Metadata

So the slicer was parked in Metadata, not coupled to it. Its seam was already
a pure function: text in, line range in, text out.

The moved code is byte-for-byte unchanged (1,827 lines verified by diff
against d67bfa42), and the surviving Metadata file is exactly main minus the
moved region. The only edits are the enclosing type, the two consumer call
sites, and two usings that the move left dead.

The library is named for what it delivers and not more: it isolates one
member's text given a PDB line range. It is not a C# parser, has no syntax
tree, and is heuristic at the edges, as its seven pinned _KnownGap tests say.

Tests move with the code, into DotnetInspector.CSharpBodySlicer.Tests.

Corpus note: AuthoredSourceValidityTests.SliceCorpus() scans every assembly in
its own output directory, so moving it changed the corpus from 4,924 slices to
3,757 (-24%). The cause is that the corpus included the test assembly itself,
and ILInspector.Metadata.Tests.dll (989 tests) is much larger than
DotnetInspector.CSharpBodySlicer.Tests.dll (244). Four small metadata-only
fixtures also dropped out; DiffAsmFixtures.Target and Decompiler.Fixtures.
NewUnsafe are referenced explicitly because the corpus needs them. The floor
the suite enforces is 1,000 slices and every rate ceiling still passes, but the
reduction is real and is recorded here rather than left to be discovered.

Slicer 244, Metadata 989 (1233 - 244 moved), Services 319, CLI 2353.

Copilot-Session: bc2613de-a721-4efd-8390-56db684eefd7
