+++ src/DotnetInspector.CSharpBodySlicer/DotnetInspector.CSharpBodySlicer.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>$(DefaultTargetFramework)</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <LangVersion>preview</LangVersion>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <Authors>Richard Lander</Authors>
+    <Description>Isolates a single member's text from a C# source file, given the line range a portable PDB reports for it. Not a C# parser: it has no syntax tree and is heuristic at the edges.</Description>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="DotnetInspector.CSharpBodySlicer.Tests" />
+  </ItemGroup>
+
+</Project>
